Made board width dynamic based on the number of tiles on mars
Specifically, based on the length of the equator. This is the last piece required for truly dynamic boards with differing sizes, since the size of hexagons is still fixed, to ensure they remain big enough to be clickable. Instead, what scales up or down is the background image of mars.